Create a lazy-array with given format and dimension

Create a directory to store lazy-array. The path must be missing. See load_lazyarray for more details

create_lazyarray(
  path,
  storage_format,
  dim,
  dimnames = NULL,
  multipart = TRUE,
  prefix = "",
  multipart_mode = 1,
  compress_level = 50L,
  meta_name = "lazyarray.meta"
)

Arguments

path
path to a local drive to store array data

storage_format
data type, choices are "double", "integer", "character", and "complex"

dim
integer vector, dimension of array, see dim

dimnames
list of vectors, names of each dimension, see dimnames

multipart
whether to split array into multiple partitions, default is true

prefix
character prefix of array partition

multipart_mode
1, or 2, mode of partition, see details.

compress_level
0 to 100, level of compression. 0 means no compression, 100 means maximum compression. For persistent data, it's recommended to set 100. Default is 50.

meta_name
header file name, default is "lazyarray.meta"

Value

A LazyArray object

Details

Lazy array stores array into hard drive, and load them on demand. It differs from other packages such as "bigmemory" that the internal reading uses multi-thread, which gains significant speed boost on solid state drives.

One lazy array contains two parts: data file(s) and a meta file. The data files can be stored in two ways: non-partitioned and partitioned.

For non-partitioned data array, the dimension is set at the creation of the array and cannot be mutable once created

For partitioned data array, there are also two partition modes, defined by `multipart_mode`. For mode 1, each partition has the same dimension size as the array, with the last dimension to be one. For example, a data with dimension c(2,3,5) partitioned with mode 1 will have each partition dimension stored with c(2,3,1). For mode 2, the last dimension will be dropped when storing each partitions.