Fitting linear models using lazyarray

Fitting linear models using lazyarray

lazylm(
  formula,
  data,
  fitted = FALSE,
  weights = NULL,
  offset = NULL,
  contrasts = NULL,
  na.action = getOption("na.action"),
  qr.tol = 1e-07,
  ...
)

Arguments

formula
an object of class formula. For variable names to be used, see 'Details'

data
a lazyarray object

fitted
whether to calculate fitted data and residuals. This may take time and memory if data is large

weights, offset, contrasts, na.action
see lm

qr.tol
the tolerance for detecting linear dependencies in the partitions of data; see qr

...
passed to chunk_map

Value

An object of class c("lazylm", "lm") or for multiple responses of class c("lazylm", "mlm").

Details

The array will be reshaped to a matrix first before fitting the linear models. A \(100 x 20 x 5\) array will be reshaped to a \(2000 x 5\) lazy matrix. The variables are the partitions of the array. If dimnames are set for the last margin index, then those will be used as variable names, otherwise lazylm automatically assign \("V1", "V2", "V3", ...\) as each partition names.

Examples

library(lazyarray)
arr <- array(rnorm(72), c(6,3,4))
arr[1,1,1] <- NA        # Allow NA to be treated
offset = rnorm(18)    # offset and weights are optional
weights = runif(18)

formula <- V1 ~ .-V2-1 + (V2 > 0)

data <- as.lazyarray(arr, type = 'file')
object <- lazylm(formula, data, weights = weights, offset = offset)

# Compare to stats::lm
dim(arr) <- c(18, 4)
lm_data <- as.data.frame(arr)
flm <- lm(formula, lm_data, weights = weights, offset = offset)

cbind(coef(object), coef(flm))

#>                   [,1]       [,2]
#> V3           0.2341853  0.2341853
#> V4          -0.1008511 -0.1008511
#> V2 > 0FALSE -0.7174550 -0.7174550
#> V2 > 0TRUE  -0.3154617 -0.3154617

cbind(resid(object), resid(flm))

#>           [,1]        [,2]
#> 2  -0.37916326 -0.37916326
#> 3  -1.54428529 -1.54428529
#> 4   0.01755074  0.01755074
#> 5   4.82317312  4.82317312
#> 6   0.30017778  0.30017778
#> 7   0.01285451  0.01285451
#> 8  -1.58389202 -1.58389202
#> 9  -2.38601485 -2.38601485
#> 10  0.22038215  0.22038215
#> 11  0.01447143  0.01447143
#> 12  0.24584364  0.24584364
#> 13  0.34766877  0.34766877
#> 14 -0.33502109 -0.33502109
#> 15  0.97765514  0.97765514
#> 16 -1.69704829 -1.69704829
#> 17  2.67366044  2.67366044
#> 18  1.01266489  1.01266489

cbind(fitted(object), fitted(flm))

#>           [,1]        [,2]
#> 2   0.32919836  0.32919836
#> 3   1.29280185  1.29280185
#> 4   0.42724637  0.42724637
#> 5  -2.06775555 -2.06775555
#> 6  -0.25364640 -0.25364640
#> 7   0.56485456  0.56485456
#> 8   1.70208689  1.70208689
#> 9   0.47429436  0.47429436
#> 10  0.64170434  0.64170434
#> 11 -0.25770817 -0.25770817
#> 12 -0.45193084 -0.45193084
#> 13 -0.32849117 -0.32849117
#> 14  0.36458184  0.36458184
#> 15 -0.42782760 -0.42782760
#> 16 -0.57706657 -0.57706657
#> 17  0.00889674  0.00889674
#> 18 -1.37388615 -1.37388615

summary(object)

#> 
#> Call:
#> lazylm(formula = formula, data = data, weights = weights, offset = offset)
#> 
#> Weighted Residuals:
#>      Min       1Q   Median       3Q      Max 
#> -1.96061 -0.33012  0.01298  0.24196  1.94063 
#> 
#> Coefficients:
#>      Estimate Std. Error t value Pr(>|t|)
#> [1,]   0.2342     0.4567   0.513    0.617
#> [2,]  -0.1009     0.5633  -0.179    0.861
#> [3,]  -0.7175     0.7469  -0.961    0.354
#> [4,]  -0.3155     0.4633  -0.681    0.508
#> 
#> Residual standard error: 1.073 on 13 degrees of freedom
#> Multiple R-squared:  0.288,	Adjusted R-squared:  0.06888 
#> F-statistic: 1.314 on 4 and 13 DF,  p-value: 0.3158
#> 

summary(flm)

#> 
#> Call:
#> lm(formula = formula, data = lm_data, weights = weights, offset = offset)
#> 
#> Weighted Residuals:
#>      Min       1Q   Median       3Q      Max 
#> -1.96061 -0.33012  0.01298  0.24196  1.94063 
#> 
#> Coefficients:
#>             Estimate Std. Error t value Pr(>|t|)
#> V3            0.2342     0.4567   0.513    0.617
#> V4           -0.1009     0.5633  -0.179    0.861
#> V2 > 0FALSE  -0.7175     0.7469  -0.961    0.354
#> V2 > 0TRUE   -0.3155     0.4633  -0.681    0.508
#> 
#> Residual standard error: 1.073 on 13 degrees of freedom
#>   (1 observation deleted due to missingness)
#> Multiple R-squared:  0.288,	Adjusted R-squared:  0.06888 
#> F-statistic: 1.314 on 4 and 13 DF,  p-value: 0.3158
#>